# backend/api/openapi/api.yaml
openapi: 3.0.3
info:
  title: School API
  description: API для платформы онлайн/оффлайн школы и помощника репетитора.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Основной шлюз API
paths:
...
  /auth/refresh:
    post:
      summary: Обновление пары токенов по refresh_token
      operationId: refreshTokens
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          description: Некорректные параметры
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Невалидный или отозванный refresh токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Получение данных текущего авторизованного пользователя
      operationId: getCurrentUser
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /clients:
    get:
      summary: Список клиентов
      operationId: listClients
      tags:
        - Clients
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
    post:
      summary: Создание клиента
      operationId: createClient
      tags:
        - Clients
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
  /clients/{id}/subscriptions:
    get:
      summary: Список абонементов
      operationId: listSubscriptions
      tags:
        - Clients
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
    post:
      summary: Создание абонемента
      operationId: createSubscription
      tags:
        - Clients
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите access_token в формате Bearer <token>
  schemas:
    Role:
      type: string
      enum:
        - student
        - teacher
        - assistant
        - owner
      example: student
    UserResponse:
      type: object
      required:
        - id
        - email
        - full_name
        - role
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        email:
          type: string
          format: email
          example: student@school.ru
        full_name:
          type: string
          example: Иван Иванов
        phone:
          type: string
          nullable: true
          example: '+79991234567'
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
          example: '2026-10-04T12:00:00Z'
    TokenPair:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: 9b2d3c1a-5f8e-4a7b-9c2d-3e4f5a6b7c8d
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Время жизни access_token в секундах
          example: 900
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: 9b2d3c1a-5f8e-4a7b-9c2d-3e4f5a6b7c8d
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input parameters
        details:
          type: array
          items:
            type: object
            additionalProperties: true
...
    ClientResponse:
      type: object
      required:
        - id
        - teacher_id
        - name
        - base_rate
        - school_percent_tag
        - created_at
      properties:
        id:
          type: string
          format: uuid
        teacher_id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
          nullable: true
        base_rate:
          type: number
        school_percent_tag:
          type: integer
        created_at:
          type: string
          format: date-time
    CreateClientRequest:
      type: object
      required:
        - name
        - base_rate
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        base_rate:
          type: number
        school_percent_tag:
          type: integer
          default: 0
    SubscriptionResponse:
      type: object
      required:
        - id
        - client_id
        - type
        - balance
        - created_at
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        type:
          type: string
        balance:
          type: number
        created_at:
          type: string
          format: date-time
    CreateSubscriptionRequest:
      type: object
      required:
        - type
        - balance
      properties:
        type:
          type: string
          enum:
            - lessons
            - hours
        balance:
          type: number
...